requirements.txt
streamlit
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-text-splitters>=0.3.0,<0.4.0
langchain-groq>=0.2.0,<0.3.0
faiss-cpu
pypdf
python-dotenv
sentence-transformers
flask
